+++ src/ObjectDescription.js
@@ -1,5 +1,8 @@
 import React, { Component } from 'react';
 
+// Styles
+import objectStyles from './objectStyles';
+
 /**
  * A short description of the object
  */
@@ -8,16 +11,16 @@ export default class ObjectDescription extends Component{
     const object = this.props.object;
     switch (typeof object){
       case 'number':
-        return (<span className="ObjectInspector-object-value-number">{object}</span>);
+        return (<span style={objectStyles.value.number}>{object}</span>);
       case 'string':
-        return (<span className="ObjectInspector-object-value-string">&quot;{object}&quot;</span>);
+        return (<span style={objectStyles.value.string}>&quot;{object}&quot;</span>);
       case 'boolean':
-        return (<span className="ObjectInspector-object-value-boolean">{String(object)}</span>);
+        return (<span style={objectStyles.value.boolean}>{String(object)}</span>);
       case 'undefined':
-        return (<span className="ObjectInspector-object-value-undefined">undefined</span>);
+        return (<span style={objectStyles.value.undefined}>undefined</span>);
       case 'object':
         if(object === null){
-          return (<span className="ObjectInspector-object-value-null">null</span>)
+          return (<span style={objectStyles.value.null}>null</span>)
         }
         if(object instanceof Date){
           return (<span>{object.toString()}</span>);
@@ -25,14 +28,14 @@ export default class ObjectDescription extends Component{
         if(Array.isArray(object)){
           return (<span>{`Array[${object.length}]`}</span>);
         }
-        return (<span className="ObjectInspector-object-value-object">Object</span>);
+        return (<span>Object</span>);
       case 'function':
         return (<span>
-                  <span className="ObjectInspector-object-value-function-keyword">function</span>
-                  <span className="ObjectInspector-object-value-function-name">&nbsp;{object.name}()</span>
+                  <span style={objectStyles.value.function.keyword}>function</span>
+                  <span style={objectStyles.value.function.name}>&nbsp;{object.name}()</span>
                 </span>);
       case 'symbol':
-        return (<span className="ObjectInspector-object-value-symbol">Symbol()</span>)
+        return (<span style={objectStyles.value.symbol}>Symbol()</span>)
       default:
         return (<span></span>);
     }